Rename facebook component
